Move lexers to correct folders
